requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
requests
pymongo
jwt
python-decouple